Fix formatting
Test: Local build
Merged-In: If78e57af0106a1cb7007b000280ec6c96b40f0d5
(cherry picked from commit 10dfad3a8659a92d3c583dccd51e1113fffd24a2)

+++ cvd/common/libs/fs/shared_fd.h
@@ -411,7 +411,7 @@ class FileInstance {
     cmsg->cmsg_level = SOL_SOCKET;
     cmsg->cmsg_type = SCM_RIGHTS;
     int* fd_array = reinterpret_cast<int*>(CMSG_DATA(cmsg));
-    for (size_t i = 0; i < SZ; ++i) {
+    for (int i = 0; i < SZ; ++i) {
       fd_array[i] = fds[i]->fd_;
     }
     return SendMsg(&msg, flags);
